Move encoding settings out of __init__.py
